@@ -24,31 +24,30 @@
     "build:cjs": "tsc --module commonjs --target es5 --outDir dist/cjs"
   },
   "devDependencies": {
-    "@types/mocha": "^7.0.2",
-    "@types/node": "^13.13.18",
-    "@typescript-eslint/eslint-plugin": "^2.34.0",
-    "@typescript-eslint/parser": "^2.34.0",
-    "c8": "^7.3.0",
-    "eslint": "^6.8.0",
-    "eslint-config-prettier": "^6.11.0",
-    "eslint-config-standard": "^14.1.1",
-    "eslint-plugin-import": "^2.22.0",
+    "@types/mocha": "^9.0.0",
+    "@types/node": "^17.0.5",
+    "@typescript-eslint/eslint-plugin": "^5.8.1",
+    "@typescript-eslint/parser": "^5.8.1",
+    "c8": "^7.10.0",
+    "eslint": "^8.5.0",
+    "eslint-config-prettier": "^8.3.0",
+    "eslint-config-standard": "^16.0.3",
+    "eslint-plugin-import": "^2.25.3",
     "eslint-plugin-node": "^11.1.0",
-    "eslint-plugin-prettier": "^3.1.4",
-    "eslint-plugin-promise": "^4.2.1",
-    "eslint-plugin-standard": "^4.0.1",
-    "husky": "^4.3.0",
-    "lint-staged": "^10.3.0",
-    "mocha": "^7.2.0",
-    "prettier": "^1.19.1",
-    "sucrase": "^3.15.0",
-    "typescript": "^4.0.2"
+    "eslint-plugin-prettier": "^4.0.0",
+    "eslint-plugin-promise": "^6.0.0",
+    "eslint-plugin-standard": "^4.1.0",
+    "husky": "^7.0.4",
+    "lint-staged": "^12.1.4",
+    "mocha": "^9.1.3",
+    "prettier": "^2.5.1",
+    "sucrase": "^3.20.3",
+    "typescript": "^4.5.4"
   },
   "bugs": {
     "url": "https://github.com/Yxwww/wyx-utils/issues"
   },
   "homepage": "https://github.com/Yxwww/wyx-utils#readme",
-  "dependencies": {},
   "keywords": [
     "utils"
   ],
@@ -59,5 +58,8 @@
   },
   "lint-staged": {
     "*.(js|ts)": "eslint --cache --fix"
+  },
+  "dependencies": {
+    "D": "^1.0.0"
   }
 }
