@@ -3,7 +3,7 @@ import { Box, id, fromNullable, noop, tryCatch, compose } from './index'
 
 describe('compose', () => {
   it('should compose function from right to left together', () => {
-    const add2 = v => v + 2
+    const add2 = (v) => v + 2
     const three = () => 3
     const threeAdd2 = compose(add2, three) as any
     assert.equal(threeAdd2(), 5)
@@ -17,10 +17,10 @@ describe('Box', () => {
   })
   it('should map correctly', () => {
     const box = Box(5)
-    assert.equal(box.map(v => v * 2).fold(id), 10)
+    assert.equal(box.map((v) => v * 2).fold(id), 10)
     box
-      .map(v => v + 1)
-      .fold(v => {
+      .map((v) => v + 1)
+      .fold((v) => {
         console.log('v', v)
       })
   })
@@ -38,13 +38,13 @@ describe('Either', () => {
   }
 
   it('should use Right when found red', () => {
-    findColor('red').fold(noop, v => {
+    findColor('red').fold(noop, (v) => {
       assert.equal('#ff0000', v)
     })
   })
 
   it('should use Left when found redd', () => {
-    findColor('redd').fold(v => {
+    findColor('redd').fold((v) => {
       assert.equal('#ff0000', v)
     }, noop)
   })
@@ -61,16 +61,16 @@ describe('Using tryCatch with Either', () => {
 
   it('should handle success case', () => {
     tryCatch(() => someSyntask(true))
-      .map(v => v + 5)
-      .fold(noop, v => {
+      .map((v) => v + 5)
+      .fold(noop, (v) => {
         assert.equal(v, 205)
       })
   })
 
   it('should handle failure', () => {
     tryCatch(() => someSyntask(false))
-      .map(v => v + 5)
-      .fold(v => {
+      .map((v) => v + 5)
+      .fold((v) => {
         assert.equal(v instanceof Error, true)
       }, noop)
   })
